--- Bitpanda2Parqet/ParqetApiResults.cs
@@ -1,77 +1,0 @@
-﻿using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Bitpanda2Parqet
-{
-    public class ParqetApiResults
-    {
-        // An object of this class contains statistical information about the via API uploaded Activities/Transactions to parqet 
-        // To Do: Check for other Api responds
-
-        public int NumberOfSuccesses { get; set; }
-        public int NumberOfDuplications { get; set; }
-        public int NumberOfUnknownCoins { get;set; }
-        public int NumberOfFailures { get; set; }
-        public int NumberOfDataSets { get; set; }
-
-        public ParqetApiResults(int numberOfSuccesses, int numberOfDuplications, int numberOfUnknownCoins, int numberOfFailures, int numberOfDataSets)
-        {
-            NumberOfSuccesses = numberOfSuccesses;
-            NumberOfDuplications = numberOfDuplications;
-            NumberOfUnknownCoins = numberOfUnknownCoins;
-            NumberOfFailures = numberOfFailures;
-            NumberOfDataSets = numberOfDataSets;
-        }
-
-        public void GetResultFromParqetResponse(string response)
-        {
-            NumberOfDataSets++;
-            JObject jsonResponse = JObject.Parse(response);
-
-            if (jsonResponse["createdActivities"].Count() == 1)
-            {
-                NumberOfSuccesses++;
-            }
-            else if (jsonResponse["notCreatedActivities"][0]["error"]["code"].ToString() == "ACTIVITY_CREATION_EXISTING_DUPLICATE")
-            {
-                NumberOfDuplications++;
-            }
-            else if (jsonResponse["notCreatedActivities"][0]["error"]["code"].ToString() == "ACTIVITY_CREATION_NO_SECURITY_DATA")
-            {
-                NumberOfUnknownCoins++;
-            }
-            else if (jsonResponse["notCreatedActivities"].Count() == 1)
-            {
-                NumberOfFailures++;
-            }
-            else
-            {
-                NumberOfFailures++;
-            }
-        }
-
-        public void SetZero()
-        {
-            NumberOfSuccesses = 0;
-            NumberOfDuplications = 0;
-            NumberOfFailures = 0;
-            NumberOfUnknownCoins = 0;
-            NumberOfDataSets = 0;
-        }
-
-        public override string ToString()
-        {
-            return "Number of Datasets: " +  NumberOfDataSets + ", where:" +
-                "\n\nNumber of Successes: " + NumberOfSuccesses +
-                "\nNumber of Duplications: " + NumberOfDuplications +
-                "\nNumber of Unknown Coins: " + NumberOfUnknownCoins +
-                "\nNumber of Fails: " + NumberOfFailures;
-        }
-
-    }
-}
